[ FEATURE ] Criação de painel de administração de usuários e middleware para admins

+++ app/Http/Controllers/UserController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserLoggedInUpdateRequest;
+use App\Http\Requests\UserRegisterRequest;
 use App\Models\User;
+use Illuminate\Database\QueryException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
@@ -15,8 +17,15 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::withoutTrashed()->get();
-        return view('users', compact('users'));
+        $is_admin = User::find(Auth::user()->id)->admin == 1;
+
+        if ($is_admin) {
+            $users = User::withTrashed()->get();
+        } else {
+            $users = User::withoutTrashed()->get();
+        }
+
+        return view('users', compact('users', 'is_admin'));
     }
 
     /**
@@ -24,15 +33,36 @@ class UserController extends Controller
      */
     public function create()
     {
-        //
+        return view('create-user');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(UserRegisterRequest $request)
     {
-        //
+        try {
+            User::create([
+                'username' => trim($request->input('username')),
+                'email' => trim($request->input('email')),
+                'password' => Hash::make(trim($request->input('password'))),
+            ]);
+        } catch (QueryException $e) {
+            // Duplicated Unique Key Error
+            if ($e->errorInfo[1] == 1062) {
+                if (str_contains($e->errorInfo[2], 'username')) {
+                    return back()
+                        ->withErrors(['username' => 'Este nome de usuário já está sendo utilizado'])
+                        ->withInput();
+                }
+
+                return back()
+                    ->withErrors(['email' => 'Este e-mail já está sendo utilizado'])
+                    ->withInput();
+            }
+        }
+
+        return redirect()->route('users.index');
     }
 
     /**
@@ -48,7 +78,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        //
+        return view('edit-user', compact('user'));
     }
 
     /**
@@ -87,16 +117,25 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        if ($user->delete()) {
-            return response("No Content", 204);
+        if ($user->admin == 1) {
+            return redirect()->back();
         }
 
-        return response("User Not Found", 404);
+        if ($user->delete()) {
+            return redirect()->back()->with('success', 'Usuário deletado');
+        }
+
+        return redirect()->back();
     }
 
     public function destroyLoggedUser(Request $request)
     {
         $user = User::find(Auth::user()->id);
+
+        if ($user->admin == 1) {
+            return response("Can't deleted logged in admin user", 400);
+        }
+
         Auth::logout();
         $request->session()->invalidate();
         $request->session()->regenerateToken();
